[enhancement] Implement Swappable models #45
