+++ openwisp_monitoring/check/classes/ping.py
@@ -4,13 +4,17 @@ from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 from jsonschema import draft7_format_checker, validate
 from jsonschema.exceptions import ValidationError as SchemaError
+from swapper import load_model
 
 from openwisp_controller.config.models import Device
 
 from ... import settings as monitoring_settings
-from ...monitoring.models import Graph, Metric, Threshold
 from .. import settings as app_settings
 from ..exceptions import OperationalError
+
+Graph = load_model('monitoring', 'Graph')
+Metric = load_model('monitoring', 'Metric')
+Threshold = load_model('monitoring', 'Threshold')
 
 
 class Ping(object):
