+++ openwisp_monitoring/device/tests/test_api.py
@@ -1,17 +1,22 @@
+from unittest import skipIf
 from unittest.mock import patch
 
 from django.contrib.auth import get_user_model
 from django.urls import reverse
+from swapper import is_swapped, load_model
 
 from openwisp_utils.tests import catch_signal
 
 from ... import settings as monitoring_settings
-from ...device.signals import device_metrics_received
-from ...monitoring.models import Graph, Metric
 from ..models import DeviceData
+from ..signals import device_metrics_received
 from . import DeviceMonitoringTestCase
 
+Graph = load_model('monitoring', 'Graph')
+Metric = load_model('monitoring', 'Metric')
 
+
+@skipIf(is_swapped('monitoring', 'Metric'), 'Running tests on sample_app')
 class TestDeviceApi(DeviceMonitoringTestCase):
     """
     Tests API (device metric collection)
