+++ openwisp_monitoring/monitoring/admin.py
@@ -1,9 +1,12 @@
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
+from swapper import load_model
 
 from openwisp_utils.admin import TimeReadonlyAdminMixin
 
-from .models import Graph, Metric, Threshold
+Graph = load_model('monitoring', 'Graph')
+Metric = load_model('monitoring', 'Metric')
+Threshold = load_model('monitoring', 'Threshold')
 
 
 class ThresholdInline(TimeReadonlyAdminMixin, admin.StackedInline):
@@ -20,18 +23,18 @@ class GraphInline(admin.StackedInline):
 
 @admin.register(Metric)
 class MetricAdmin(TimeReadonlyAdminMixin, admin.ModelAdmin):
-    list_display = ('__str__', 'created', 'modified')
+    list_display = ['__str__', 'created', 'modified']
     readonly_fields = ['is_healthy']
     search_fields = ['name']
     save_on_top = True
     inlines = [GraphInline, ThresholdInline]
-    fieldsets = (
+    fieldsets = [
         (None, {'fields': ('name', 'description', 'content_type', 'object_id',)}),
         (
             _('Advanced options'),
             {'classes': ('collapse',), 'fields': ('key', 'field_name')},
         ),
-    )
+    ]
 
     class Media:
         css = {'all': ('monitoring/css/monitoring.css',)}
@@ -40,6 +43,6 @@ class MetricAdmin(TimeReadonlyAdminMixin, admin.ModelAdmin):
 
 @admin.register(Threshold)
 class ThresholdAdmin(TimeReadonlyAdminMixin, admin.ModelAdmin):
-    list_display = ('metric', 'created', 'modified')
+    list_display = ['metric', 'created', 'modified']
     search_fields = ['name']
     save_on_top = True
