+++ tests/openwisp2/sample_device_monitoring/__init__.py
@@ -1,0 +1,3 @@
+default_app_config = (
+    'openwisp2.sample_device_monitoring.apps.SampleDeviceMonitoringConfig'
+)
