+++ tests/openwisp2/sample_device_monitoring/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from openwisp_monitoring.device.admin import DeviceAdmin  # noqa
+
+from .models import DetailsModel
+
+admin.site.register(DetailsModel)
