+++ tests/openwisp2/sample_device_monitoring/apps.py
@@ -1,0 +1,6 @@
+from openwisp_monitoring.device.apps import DeviceMonitoringConfig
+
+
+class SampleDeviceMonitoringConfig(DeviceMonitoringConfig):
+    name = 'openwisp2.sample_device_monitoring'
+    label = 'sample_device_monitoring'
