+++ tests/openwisp2/sample_monitoring/apps.py
@@ -1,0 +1,6 @@
+from openwisp_monitoring.monitoring.apps import MonitoringConfig
+
+
+class SampleMonitoringConfig(MonitoringConfig):
+    name = 'openwisp2.sample_monitoring'
+    label = 'sample_monitoring'
